Support message-reception in library and repl
Turns out I was doing three things wrong:
  - Inverting the receiver IQ rather than just the transceiver
  - I implemented the errata note in the main spec
    (clear the top bit of 0x31) without the accompanying extra
    steps from the actual errata doc (re-set the values for 0x2f, 0x30
    after doing so) and so my code was breaking while radiolib's was
    fine (as they do not implement the errata note)
  - Casting to (from?) the irqs bitfield was losing data
I should still move to listening on the dio line, or at least polling on
the status. Moreover, the docs allege I should be using RxContinuous,
though I don't know whether or not that will work with the anticipated
single-radio architecture of the chat system. Still, anything is better
than the current method of hardcoding a wait-time!