Define radio interface for libbcp
I'm unsure about whether I actually want an abstraction layer here, as I
do right now: on the plus side it means I could theoretically run bcp on
a different physical platform, but on the minus side it makes it
difficult to do anything which cannot be abstracted across radios, e.g.
changing LoRa knobs (viz. cr, sf, bw, freq) on the fly.
Adding it in for now because it's easier to tear it out later than build
it in later.