Implement and adopt continuous message reception mode
Despite the name, we still use it to receive a single message at a time
(though we could handle multiple): the difference is at the hardware
level, where in "Single receive mode" the SX1276 will receive a single
message before dropping out of reception mode automatically, while in
"Continuous receive mode" it will return to reception mode. To some
extent this is not desirable: it means that we need to filter for
multiple packets ourselves. However, it is the mode recommended by the
SX1276 datasheet, and moreover using it (plus some fixes) has produced a
much more functional result than what we were previously getting using
the single-reception mode.
There is a caveat: it seems that there's a hardware issue wherein
resetting any interrupt while the RxDone bit is set following a
continuous-reception mode will result in the whole device switching to
and seemingly locking to FSK frequency-synthesis mode! This is obviously
bad, but I was able to work around it by masking off the RxDone bit and
just watching the ValidHeader bit instead. However, this has the side
effect of making it impossible to distinguish a partially-received
packet from a fully-received one. As such, it may be worth looking into
this single-reception mode again in the future.
FWIW radiolib uses single-reception mode as well.