libsx1276: properly calculate time-on-air for TX/RX
This is straight out of the datasheets -- finding Ts was a pain since
for whatever reason they decided not to define it in the 127x sheet so I
had to look elsewhere.
Not entirely happy that I have to add a ~75ms fudge factor (perhaps this
is because we don't/can't wait for the RxDone IRQ? but experimentally it
seems like the issue is on the Tx side not going long enough, which
wouldn't line up), but it's still noticably snappier than before.

Woohoo!